Fix: Fee token implementation separated from the wrapper
